CometModel class + run download in setup
