+++ setup.py
@@ -1,7 +1,18 @@
+import os
 import setuptools
+
+from setuptools.command.install import install
 
 with open("README.md", "r") as fh:
     long_description = fh.read()
+
+
+class MyInstall(install):
+    def run(self):
+        install.run(self)
+        script = os.path.join(os.getcwd(), "setup/download.sh")
+        os.system(f"bash {script}")
+
 
 setuptools.setup(
     name="comet-commonsense", # Replace with your own username
@@ -18,4 +29,5 @@ setuptools.setup(
         "Operating System :: OS Independent",
     ],
     python_requires='>=3.6',
+    cmdclass={'install': MyInstall}
 )
