+++ src/evaluate.py
@@ -1,11 +1,12 @@
 import tqdm
-import torch
 import logging
 import argparse
 
+from collections import defaultdict
 from nltk.translate.bleu_score import sentence_bleu, SmoothingFunction
 
-from src.common import init_model, generate_ending, get_atomic_categories, load_atomic_data_for_eval
+from src.atomic import load_atomic_data
+from src.comet_model import PretrainedCometModel
 
 logging.basicConfig(
     format = '%(asctime)s - %(levelname)s - %(name)s -   %(message)s',
@@ -32,14 +33,16 @@ def main():
     args = parser.parse_args()
     logger.debug(args)
 
-    device = torch.device(f"cuda:{args.device}" if torch.cuda.is_available() and args.device != "cpu" else "cpu")
-    logger.debug(f"Initializing {args.device}")
-
-    tokenizer, model = init_model(args.model_name_or_path, device=device, do_lower_case=args.do_lower_case)
-    categories = get_atomic_categories()
+    comet_model = PretrainedCometModel(args.model_name_or_path, device=args.device, do_lower_case=args.do_lower_case)
 
     logger.info(f"Loading ATOMIC examples from {args.in_file}")
-    examples = load_atomic_data_for_eval(args.in_file, categories, tokenizer)
+    examples = load_atomic_data(args.in_file, comet_model.categories)
+
+    examples_by_category = defaultdict(lambda: defaultdict(list))
+
+    for e1, curr_relations in examples.items():
+        for cat, e2s in curr_relations.items():
+            examples_by_category[cat][e1.replace("<blank>", "___")] = [f"{e2} <eos>" for e2 in e2s]
 
     for generation_arg in [{"k": 1}, {"k": 5}, {"k": 10},
                            {"num_beams": 10, "k": 10},
@@ -50,19 +53,19 @@ def main():
         total_bl = {}
         total_count = {}
 
-        for category, curr_examples in examples.items():
+        for category, curr_examples in examples_by_category.items():
             total_bl[category] = 0
             total_count[category] = 0
 
-            for input_event, (input_tokens, list_of_refs) in tqdm.tqdm(curr_examples.items()):
+            for input_event, list_of_refs in tqdm.tqdm(curr_examples.items()):
                 list_of_refs = [ref for ref in list_of_refs if ref != "none"]
                 if len(list_of_refs) == 0:
                     continue
 
-                list_of_refs = [tokenizer.tokenize(ref) for ref in list_of_refs]
+                list_of_refs = [comet_model.tokenizer.tokenize(ref) for ref in list_of_refs]
 
-                system_outputs = generate_ending(
-                    model, tokenizer, input_tokens, device, length=args.max_length, num_samples=args.num_samples,
+                system_outputs = comet_model.predict(
+                    input_event, category, length=args.max_length, num_samples=args.num_samples,
                     return_tokenized=True, **generation_arg)
 
                 bleu_scores = [sentence_bleu(
