+++ src/predict.py
@@ -1,14 +1,13 @@
 import json
 import tqdm
-import torch
 import logging
 import argparse
 
-from src.common import init_model, generate_ending, get_atomic_categories, load_atomic_data_for_generation
+from src.atomic import load_atomic_data
+from src.comet_model import PretrainedCometModel
 
 logging.basicConfig(
-    format = '%(asctime)s - %(levelname)s - %(name)s -   %(message)s',
-    datefmt = '%m/%d/%Y %H:%M:%S', level = logging.INFO)
+    format='%(asctime)s - %(levelname)s - %(name)s -   %(message)s', datefmt='%m/%d/%Y %H:%M:%S', level=logging.INFO)
 
 logger = logging.getLogger(__name__)
 
@@ -39,22 +38,19 @@ def main():
     #         sum([1 if i > 0 else 0 for i in [args.k, args.p, args.num_beams]]) > 1):
     #     raise ValueError("Exactly one of p, k, or num_beams should be set to a non-zero value.")
 
-    device = torch.device(f"cuda:{args.device}" if torch.cuda.is_available() and args.device != "cpu" else "cpu")
-    logger.debug(f"Initializing {args.device}")
-
-    tokenizer, model = init_model(args.model_name_or_path, device=device, do_lower_case=args.do_lower_case)
-    categories = get_atomic_categories()
+    comet_model = PretrainedCometModel(args.model_name_or_path, device=args.device, do_lower_case=args.do_lower_case)
 
     logger.info(f"Loading ATOMIC examples from {args.in_file}")
-    examples = load_atomic_data_for_generation(args.in_file, categories, tokenizer)
+    examples = load_atomic_data(args.in_file, comet_model.categories)
+    examples = {e1: list(e1_relations.keys()) for e1, e1_relations in examples.items()}
 
     with open(args.out_file, "w", encoding="utf-8") as f_out:
         for input_event, curr_categories in tqdm.tqdm(examples.items()):
             example = {"input": input_event}
-            for category, tokens in curr_categories.items():
-                example[category] = generate_ending(
-                    model, tokenizer, tokens, device, p=args.p, k=args.k,
-                    num_beams=args.num_beams, length=args.max_length, num_samples=args.num_samples)
+            for category in curr_categories:
+                example[category] = comet_model.predict(
+                    input_event, category, p=args.p, k=args.k, num_beams=args.num_beams, length=args.max_length,
+                    num_samples=args.num_samples)
 
             f_out.write(json.dumps(example) + "\n")
